+++ crates/aranya-policy-module/src/data.rs
@@ -5,7 +5,7 @@ use core::fmt::{self, Display};
 
 pub use aranya_crypto::Id;
 use aranya_crypto::{DeviceId, EncryptionKeyId};
-use aranya_policy_ast::VType;
+use aranya_policy_ast::{Identifier, Text, VType};
 use serde::{Deserialize, Serialize};
 
 use super::ffi::Type;
@@ -25,7 +25,7 @@ pub enum ValueConversionError {
     },
     /// A struct conversion found a field mismatch between types
     #[error("invalid struct member `{0}`")]
-    InvalidStructMember(String),
+    InvalidStructMember(Identifier),
     /// The target type does not have sufficient range to represent this Value
     #[error("value out of range")]
     OutOfRange,
@@ -64,8 +64,7 @@ macro_rules! impl_typed {
     };
 }
 
-impl_typed!(String => String);
-impl_typed!(&str => String);
+impl_typed!(Text => String);
 
 impl_typed!(Vec<u8> => Bytes);
 impl_typed!(&[u8] => Bytes);
@@ -89,7 +88,7 @@ impl_typed!(EncryptionKeyId => Id);
 impl_typed!(DeviceId => Id);
 
 impl<T: Typed> Typed for Option<T> {
-    const TYPE: Type<'static> = Type::Optional(&T::TYPE);
+    const TYPE: Type<'static> = Type::Optional(const { &T::TYPE });
 }
 
 /// All of the value types allowed in the VM
@@ -100,7 +99,7 @@ pub enum Value {
     /// Boolean
     Bool(bool),
     /// String (UTF-8)
-    String(String),
+    String(Text),
     /// Bytes
     Bytes(Vec<u8>),
     /// Struct
@@ -110,7 +109,9 @@ pub enum Value {
     /// A unique identifier.
     Id(Id),
     /// Enumeration value
-    Enum(String, i64),
+    Enum(Identifier, i64),
+    /// Textual Identifier (name)
+    Identifier(Identifier),
     /// Empty optional value
     None,
 }
@@ -176,6 +177,7 @@ impl Value {
             Value::Fact(f) => format!("Fact {}", f.name),
             Value::Id(_) => String::from("Id"),
             Value::Enum(name, _) => format!("Enum {}", name),
+            Value::Identifier(_) => String::from("Identifier"),
             Value::None => String::from("None"),
         }
     }
@@ -218,15 +220,15 @@ impl From<bool> for Value {
     }
 }
 
-impl From<&str> for Value {
-    fn from(value: &str) -> Self {
-        Value::String(value.to_owned())
+impl From<Text> for Value {
+    fn from(value: Text) -> Self {
+        Value::String(value)
     }
 }
 
-impl From<String> for Value {
-    fn from(value: String) -> Self {
-        Value::String(value)
+impl From<Identifier> for Value {
+    fn from(value: Identifier) -> Self {
+        Value::Identifier(value)
     }
 }
 
@@ -302,7 +304,7 @@ impl TryFrom<Value> for bool {
     }
 }
 
-impl TryFrom<Value> for String {
+impl TryFrom<Value> for Text {
     type Error = ValueConversionError;
 
     fn try_from(value: Value) -> Result<Self, Self::Error> {
@@ -312,8 +314,23 @@ impl TryFrom<Value> for String {
         Err(ValueConversionError::invalid_type(
             "String",
             value.type_name(),
-            "Value -> String",
+            "Value -> Text",
         ))
+    }
+}
+
+impl TryFrom<Value> for Identifier {
+    type Error = ValueConversionError;
+
+    fn try_from(value: Value) -> Result<Self, Self::Error> {
+        let Value::Identifier(text) = value else {
+            return Err(ValueConversionError::invalid_type(
+                "Identifier",
+                value.type_name(),
+                "Value -> Identifier",
+            ));
+        };
+        Ok(text)
     }
 }
 
@@ -438,20 +455,6 @@ impl TryAsMut<bool> for Value {
     }
 }
 
-impl TryAsMut<str> for Value {
-    type Error = ValueConversionError;
-    fn try_as_mut(&mut self) -> Result<&mut str, Self::Error> {
-        if let Self::String(s) = self {
-            return Ok(s);
-        }
-        Err(ValueConversionError::invalid_type(
-            "String",
-            self.type_name(),
-            "Value -> String",
-        ))
-    }
-}
-
 impl TryAsMut<[u8]> for Value {
     type Error = ValueConversionError;
     fn try_as_mut(&mut self) -> Result<&mut [u8], Self::Error> {
@@ -511,6 +514,7 @@ impl Display for Value {
             Value::Fact(fa) => fa.fmt(f),
             Value::Id(id) => id.fmt(f),
             Value::Enum(name, value) => write!(f, "{name}::{value}"),
+            Value::Identifier(name) => write!(f, "{name}"),
             Value::None => write!(f, "None"),
         }
     }
@@ -526,11 +530,11 @@ pub enum HashableValue {
     /// A bool.
     Bool(bool),
     /// A string.
-    String(String),
+    String(Text),
     /// A unique identifier.
     Id(Id),
     /// Enum
-    Enum(String, i64),
+    Enum(Identifier, i64),
 }
 
 impl HashableValue {
@@ -591,16 +595,16 @@ impl Display for HashableValue {
 #[cfg_attr(feature = "proptest", derive(proptest_derive::Arbitrary))]
 pub struct FactKey {
     /// key name
-    pub identifier: String,
+    pub identifier: Identifier,
     /// key value
     pub value: HashableValue,
 }
 
 impl FactKey {
     /// Creates a new fact key.
-    pub fn new(name: &str, value: HashableValue) -> Self {
+    pub fn new(name: Identifier, value: HashableValue) -> Self {
         Self {
-            identifier: String::from(name),
+            identifier: name,
             value,
         }
     }
@@ -616,16 +620,16 @@ impl Display for FactKey {
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct FactValue {
     /// value name
-    pub identifier: String,
+    pub identifier: Identifier,
     /// value
     pub value: Value,
 }
 
 impl FactValue {
     /// Creates a new fact value.
-    pub fn new(name: &str, value: Value) -> Self {
+    pub fn new(name: Identifier, value: Value) -> Self {
         Self {
-            identifier: String::from(name),
+            identifier: name,
             value,
         }
     }
@@ -647,21 +651,21 @@ pub type FactValueList = Vec<FactValue>;
 /// Technically identical to a FactValue but separate to distinguish
 /// usage.
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct KVPair(String, Value);
+pub struct KVPair(Identifier, Value);
 
 impl KVPair {
     /// Creates a key-value pair.
-    pub fn new(key: &str, value: Value) -> KVPair {
-        KVPair(key.to_owned(), value)
+    pub fn new(key: Identifier, value: Value) -> KVPair {
+        KVPair(key, value)
     }
 
     /// Creates a key-value pair with an integer value.
-    pub fn new_int(key: &str, value: i64) -> KVPair {
-        KVPair(key.to_owned(), Value::Int(value))
+    pub fn new_int(key: Identifier, value: i64) -> KVPair {
+        KVPair(key, Value::Int(value))
     }
 
     /// Returns the key half of the key-value pair.
-    pub fn key(&self) -> &str {
+    pub fn key(&self) -> &Identifier {
         &self.0
     }
 
@@ -677,13 +681,13 @@ impl Display for KVPair {
     }
 }
 
-impl From<KVPair> for (String, Value) {
+impl From<KVPair> for (Identifier, Value) {
     fn from(kv: KVPair) -> Self {
         (kv.0, kv.1)
     }
 }
 
-impl From<&KVPair> for (String, Value) {
+impl From<&KVPair> for (Identifier, Value) {
     fn from(value: &KVPair) -> Self {
         (value.0.clone(), value.1.clone())
     }
@@ -705,7 +709,7 @@ impl From<FactValue> for KVPair {
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct Fact {
     /// The name of the fact
-    pub name: String,
+    pub name: Identifier,
     /// The keys of the fact
     pub keys: FactKeyList,
     /// The values of the fact
@@ -714,7 +718,7 @@ pub struct Fact {
 
 impl Fact {
     /// Creates a fact.
-    pub fn new(name: String) -> Fact {
+    pub fn new(name: Identifier) -> Fact {
         Fact {
             name,
             keys: vec![],
@@ -723,18 +727,18 @@ impl Fact {
     }
 
     /// Sets the fact's key.
-    pub fn set_key<V>(&mut self, name: String, value: V)
+    pub fn set_key<V>(&mut self, name: Identifier, value: V)
     where
         V: Into<HashableValue>,
     {
         match self.keys.iter_mut().find(|e| e.identifier == name) {
-            None => self.keys.push(FactKey::new(&name, value.into())),
+            None => self.keys.push(FactKey::new(name, value.into())),
             Some(e) => e.value = value.into(),
         }
     }
 
     /// Sets the fact's value.
-    pub fn set_value<V>(&mut self, name: String, value: V)
+    pub fn set_value<V>(&mut self, name: Identifier, value: V)
     where
         V: Into<Value>,
     {
@@ -784,31 +788,21 @@ impl Display for Fact {
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct Struct {
     /// The name of the struct
-    pub name: String,
+    pub name: Identifier,
     /// the fields of the struct
-    pub fields: BTreeMap<String, Value>,
+    pub fields: BTreeMap<Identifier, Value>,
 }
 
 impl Struct {
     /// Creates a struct.
-    pub fn new(name: &str, fields: impl IntoIterator<Item = impl Into<(String, Value)>>) -> Struct {
+    pub fn new(
+        name: Identifier,
+        fields: impl IntoIterator<Item = impl Into<(Identifier, Value)>>,
+    ) -> Struct {
         Struct {
-            name: name.to_owned(),
+            name,
             fields: fields.into_iter().map(|p| p.into()).collect(),
         }
-    }
-}
-
-impl From<Struct> for (String, Vec<KVPair>) {
-    fn from(value: Struct) -> Self {
-        (
-            value.name,
-            value
-                .fields
-                .into_iter()
-                .map(|(k, v)| KVPair(k, v))
-                .collect(),
-        )
     }
 }